fix(conflict_manager_storage): replace index_data in uint64

+++ storage/inc/com/centreon/broker/storage/conflict_manager.hh
@@ -107,7 +107,7 @@ class conflict_manager {
 
   struct index_info {
     std::string host_name;
-    uint32_t index_id;
+    uint64_t index_id;
     bool locked;
     uint32_t rrd_retention;
     std::string service_description;
